refactor: 시내버스 로직 분리
